The real deal
- Finished entire code
- Finished implementing my touch and personality into the code (It's a game after all)
- Fixed all bugs
I'm very proud of this and really enjoyed the process of doing this. I noticed the more I planned to add more features, the more things and little tiny problems started popping out due to the large amount of pieces all moving together. In the end I managed to make it work and tested it multiple times.